move sorted_uniforms to func.rs
